Refactor all Match Service, so now is more simple and autonomous when incrementing reps; CLub & Athlete Controllers implemented but needing manual tests; Match Controller TODO.

+++ back-end/repo-jpa/src/main/kotlin/com/caliarena/repo/jpa/routine/ExerciseRepositoryJpa.kt
@@ -3,6 +3,6 @@ package com.caliarena.repo.jpa.routine
 import com.caliarena.repo.entities.routine.ExerciseEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface ExerciseRepositoryJpa : JpaRepository<ExerciseEntity, Long> {
+interface ExerciseRepositoryJpa : JpaRepository<ExerciseEntity, Int> {
     fun findExercisesByRoutineId(routineId: Int): List<ExerciseEntity>
 }
